Criação da classe Stock, adição do tratamento de excepções e outras melhorias

+++ src/main/java/com/example/sistemagestao/services/IngredientService.java
@@ -8,6 +8,8 @@ import com.example.sistemagestao.dto.IngredientRequestDTO;
 import com.example.sistemagestao.dto.IngredientResponseDTO;
 import com.example.sistemagestao.repositories.IngredientRepository;
 import com.example.sistemagestao.repositories.RecipeIngredientsRepository;
+import com.example.sistemagestao.repositories.StockRepository;
+import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -24,7 +26,9 @@ public class IngredientService {
     @Autowired
     private IngredientRepository ingredientRepository;
     @Autowired
-    private RecipeIngredientsRepository recipeIngredientsRepository;
+    private StockRepository stockRepository;
+    @Autowired
+    private StockService stockService;
 
     public List<IngredientResponseDTO> getAll() {
         return ingredientRepository.findAllByOrderByNameAsc()
@@ -40,15 +44,27 @@ public class IngredientService {
 
     public IngredientResponseDTO getById(Long id) {
         return new IngredientResponseDTO(ingredientRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Ingrediente com ID " + id + " não encontrado")));
+                .orElseThrow(() -> new EntityNotFoundException("Ingrediente não encontrado.")));
+    }
+
+    public List<IngredientResponseDTO> getAllWithoutStock() {
+        return ingredientRepository.findAllWithNoStockInAnyBakery()
+                .stream().map(IngredientResponseDTO::new)
+                .toList();
     }
 
     @Transactional
     public void add(IngredientRequestDTO data) {
         MeasurentUnits unit = MeasurentUnits.findByDescription(data.unitDescription());
 
+        if (ingredientRepository.existsByName(data.name())) {
+            throw new EntityExistsException("Já existe um Ingrediente com esse nome.");
+        }
+
         Ingredient ingredientData = new Ingredient(data, unit);
         ingredientRepository.save(ingredientData);
+
+        stockService.initializeStock(ingredientData.getId());
     }
 
     @Transactional
@@ -58,7 +74,7 @@ public class IngredientService {
             unit = MeasurentUnits.findByDescription(newData.unitDescription());
 
         Ingredient ingredient = ingredientRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Ingrediente com ID " + id + " não encontrado"));
+                .orElseThrow(() -> new EntityNotFoundException("Ingrediente não encontrado."));
 
         ingredient.updateIngredient(newData, unit);
         ingredientRepository.save(ingredient);
@@ -67,10 +83,13 @@ public class IngredientService {
     @Transactional
     public void deleteById(Long id) {
         Ingredient ingredient = ingredientRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Ingrediente com ID " + id + " não encontrado"));
+                .orElseThrow(() -> new EntityNotFoundException("Ingrediente não encontrado."));
 
-        List<RecipeIngredient> relatedIngredients = recipeIngredientsRepository.findAllByIngredientId(id);
-        recipeIngredientsRepository.deleteAll(relatedIngredients);
+        boolean haveStock = stockRepository.existsByIngredientIdAndQuantityGreaterThan(id, 0.0);
+
+        if (haveStock) {
+            throw new IllegalStateException("Ingrediente com stock em pelo menos uma pastelaria.");
+        }
 
         ingredientRepository.delete(ingredient);
     }
